+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 **Breaking changes**
 
-- Changed the I18n token `euiDataGridToolbar.fullScreenButton` to `euiFullScreenSelector.fullScreenButton`, and its text to `Enter fullscreen` (with no space) ([#5680](https://github.com/elastic/eui/pull/5680))
-- Changed the I18n token `euiDataGridToolbar.fullScreenButtonActive` to `euiFullScreenSelector.fullScreenButtonActive`, and its text to `Exit fullscreen` (with no space) ([#5680](https://github.com/elastic/eui/pull/5680))
+- Changed the I18n token `euiDataGridToolbar.fullScreenButton` to `euiFullscreenSelector.fullscreenButton`, and its text to `Enter fullscreen` (with no space) ([#5680](https://github.com/elastic/eui/pull/5680))
+- Changed the I18n token `euiDataGridToolbar.fullScreenButtonActive` to `euiFullscreenSelector.fullscreenButtonActive`, and its text to `Exit fullscreen` (with no space) ([#5680](https://github.com/elastic/eui/pull/5680))
 
 ## [`51.0.0`](https://github.com/elastic/eui/tree/v51.0.0)